backend/requirements.txt
Django
djangorestframework
psycopg2
python-dotenv
django-cors-headers
djangorestframework-simplejwt
nanoid
drf-spectacular
drf-yasg
django-filter
Pillow
requests
vnstock
gunicorn
apscheduler
pandas
bcrypt
boto3
langchain
langchain-openai
langchain-community
python-docx
pgvector
openai